Refactor file structure and naming conventions.
Renamed files and updated imports across the codebase to align with a more organized structure and consistent naming conventions. Changes include adding context-specific prefixes and grouping related modules under appropriate directories.
