+++ src/auth/auth-controller.js
@@ -1,6 +1,6 @@
 import jwt from 'jsonwebtoken';
 import * as argon2 from "argon2";
-import UserController from "../user/controller.js";
+import UserController from "../user/user-controller.js";
 import * as mailer from "../services/mailer/mailer-service.js";
 
 class AuthController extends UserController {
