+++ src/auth/auth-routes.js
@@ -1,6 +1,6 @@
 import express from "express";
 const router = express.Router();
-import authController from './controller.js';
+import authController from './auth-controller.js';
 
 router.post('/register',
     authController.setAccessPolicy.bind(authController),
