+++ src/services/scheduler/scheduler-service.js
@@ -1,5 +1,5 @@
 import cron from 'node-cron';
-import { notifyUpcomingEventTask } from './tasks/notify-upcoming-event-task.js';
+import { notifyUpcomingEventTask } from './tasks/event/notify-upcoming-event-task.js';
 
 cron.schedule('* * * * *', async () => {
     console.group('Run scheduler: * * * * * (every minute)');
